requirements.txt
appdirs==1.4.4
asgiref==3.3.1
boto3==1.17.34
botocore==1.20.34
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
distlib==0.3.1
Django==3.1.7
django-storages==1.11.1
filelock==3.0.12
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
jmespath==0.10.0
jwcrypto==0.8
jws==0.1.3
oauth2client==3.0.0
Pillow==8.1.1
pipenv==2020.11.15
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2021.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.2
virtualenv-clone==0.5.4
